remove the need for lasa_GP
